move category to own component
